# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Hospital Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/appointments/{appointment}": {
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update an appointment",
                "operationId": "6d90dc585178baa81c3ab129afe924ec",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Delete an appointment",
                "operationId": "d2e4aef8da50e5d82f441ed65ea4700d",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Appointment deleted"
                    }
                }
            }
        },
...
        "/doctors/{doctor}": {
            "put": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Update a doctor",
                "operationId": "8567b03dce7c69a624a0896b5a1d8a4e",
                "parameters": [
                    {
                        "name": "doctor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Delete a doctor",
                "operationId": "4aef2104bf44af14814420dde8ee9cca",
                "parameters": [
                    {
                        "name": "doctor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Doctor deleted"
                    }
                }
            }
        },
...
        "/invoices/{invoice}": {
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Update an invoice",
                "operationId": "a40e07f5b2e5800f3ae9730319040d51",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Delete an invoice",
                "operationId": "1a2cf0dde9ee2180fbdaf078b4329597",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Invoice deleted"
                    }
                }
            }
        },
...
        "/medications/{medication}": {
            "put": {
                "tags": [
                    "Medications"
                ],
                "summary": "Update a medication",
                "operationId": "ef60cc2bbc9a140bf9aecb4137b420cc",
                "parameters": [
                    {
                        "name": "medication",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medication updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Medications"
                ],
                "summary": "Delete a medication",
                "operationId": "3187b422dd4432a8f840b4a9e77f80d3",
                "parameters": [
                    {
                        "name": "medication",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Medication deleted"
                    }
                }
            }
        },
...
        "/patients/{patient}": {
            "put": {
                "tags": [
                    "Patients"
                ],
                "summary": "Update a patient",
                "operationId": "77c375326210f241e7e72e2e96795148",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Patients"
                ],
                "summary": "Delete a patient",
                "operationId": "1ef1fdeaafba09c2635d2abffc2291de",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Patient deleted"
                    }
                }
            }
        },
...
        "/payments/{payment}": {
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Update a payment",
                "operationId": "5e6ce68c765bfba967264070023c3122",
                "parameters": [
                    {
                        "name": "payment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Delete a payment",
                "operationId": "10dde42bcf9365c33158aa3c769af513",
                "parameters": [
                    {
                        "name": "payment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Payment deleted"
                    }
                }
            }
        },
...
        "/prescriptions/{prescription}": {
            "put": {
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Update a prescription",
                "operationId": "fbb6a319782fc9b0defdb4d6dc71d2c7",
                "parameters": [
                    {
                        "name": "prescription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prescription updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Delete a prescription",
                "operationId": "f612936e5058ab2da31227ecab9d85d6",
                "parameters": [
                    {
                        "name": "prescription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Prescription deleted"
                    }
                }
            }
        },
...
        "/rooms/{room}": {
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update a room",
                "operationId": "92aa8019b40cb4324223ff1dbec0d091",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room",
                "operationId": "9379b2428c4e684a05dc9a71d9ee894e",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Room deleted"
                    }
                }
            }
        },
...
        "/treatments/{treatment}": {
            "put": {
                "tags": [
                    "Treatments"
                ],
                "summary": "Update a treatment",
                "operationId": "6a9823a9e505f881085f7511fada3335",
                "parameters": [
                    {
                        "name": "treatment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Treatment updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Treatments"
                ],
                "summary": "Delete a treatment",
                "operationId": "18df29b45b34bdf4565f900d595f48c4",
                "parameters": [
                    {
                        "name": "treatment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Treatment deleted"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}